+++ search/quiescence_search.hpp
@@ -7,7 +7,7 @@
 #include "../chess_board/board.hpp"
 #include "../move_generation/move_list.hpp"
 #include "../move_generation/move_generator.hpp"
-#include "../evaluation/evaluation.hpp"
+#include "../executioner/makemove.hpp"
 
 template <Color color>
 std::int16_t quiescence_search(board & chessboard, search_data & data, std::int16_t alpha, std::int16_t beta) {
